Add support for processing WAN replication messages from 3.x members
Adds support on PASSIVE (target) cluster to be able to deserialize and
process WAN replication events from a 3.x ACTIVE (source) cluster.
