+++ hazelcast/src/main/java/com/hazelcast/internal/serialization/impl/AbstractSerializationService.java
@@ -17,6 +17,7 @@
 package com.hazelcast.internal.serialization.impl;
 
 import com.hazelcast.core.ManagedContext;
+import com.hazelcast.internal.compatibility.serialization.impl.CompatibilitySerializationConstants;
 import com.hazelcast.internal.nio.BufferObjectDataInput;
 import com.hazelcast.internal.nio.BufferObjectDataOutput;
 import com.hazelcast.internal.serialization.Data;
@@ -73,9 +74,8 @@ public abstract class AbstractSerializationService implements InternalSerializat
     protected SerializerAdapter javaSerializerAdapter;
     protected SerializerAdapter javaExternalizableAdapter;
 
-    private final IdentityHashMap<Class, SerializerAdapter> constantTypesMap = new IdentityHashMap<Class, SerializerAdapter>(
-            CONSTANT_SERIALIZERS_LENGTH);
-    private final SerializerAdapter[] constantTypeIds = new SerializerAdapter[CONSTANT_SERIALIZERS_LENGTH];
+    private final IdentityHashMap<Class, SerializerAdapter> constantTypesMap;
+    private final SerializerAdapter[] constantTypeIds;
     private final ConcurrentMap<Class, SerializerAdapter> typeMap = new ConcurrentHashMap<Class, SerializerAdapter>();
     private final ConcurrentMap<Integer, SerializerAdapter> idMap = new ConcurrentHashMap<Integer, SerializerAdapter>();
     private final AtomicReference<SerializerAdapter> global = new AtomicReference<SerializerAdapter>();
@@ -100,6 +100,12 @@ public abstract class AbstractSerializationService implements InternalSerializat
         this.bufferPoolThreadLocal = new BufferPoolThreadLocal(this, builder.bufferPoolFactory,
                 builder.notActiveExceptionSupplier);
         this.nullSerializerAdapter = createSerializerAdapter(new ConstantSerializers.NullSerializer());
+        this.constantTypesMap = new IdentityHashMap<>(builder.isCompatibility
+                ? CompatibilitySerializationConstants.CONSTANT_SERIALIZERS_LENGTH
+                : SerializationConstants.CONSTANT_SERIALIZERS_LENGTH);
+        this.constantTypeIds = new SerializerAdapter[builder.isCompatibility
+                ? CompatibilitySerializationConstants.CONSTANT_SERIALIZERS_LENGTH
+                : SerializationConstants.CONSTANT_SERIALIZERS_LENGTH];
     }
 
     // used by jet
@@ -114,6 +120,8 @@ public abstract class AbstractSerializationService implements InternalSerializat
         this.bufferPoolThreadLocal = new BufferPoolThreadLocal(this, new BufferPoolFactoryImpl(),
                 prototype.notActiveExceptionSupplier);
         this.nullSerializerAdapter = prototype.nullSerializerAdapter;
+        this.constantTypesMap = new IdentityHashMap<>(prototype.constantTypesMap.size());
+        this.constantTypeIds = new SerializerAdapter[prototype.constantTypeIds.length];
     }
 
     //region Serialization Service
@@ -459,7 +467,7 @@ public abstract class AbstractSerializationService implements InternalSerializat
     public SerializerAdapter serializerFor(final int typeId) {
         if (typeId <= 0) {
             final int index = indexForDefaultType(typeId);
-            if (index < CONSTANT_SERIALIZERS_LENGTH) {
+            if (index < constantTypeIds.length) {
                 return constantTypeIds[index];
             }
         }
@@ -590,6 +598,7 @@ public abstract class AbstractSerializationService implements InternalSerializat
         private int initialOutputBufferSize;
         private BufferPoolFactory bufferPoolFactory;
         private Supplier<RuntimeException> notActiveExceptionSupplier;
+        private boolean isCompatibility;
 
         protected Builder() {
         }
@@ -638,6 +647,27 @@ public abstract class AbstractSerializationService implements InternalSerializat
         public final T withNotActiveExceptionSupplier(Supplier<RuntimeException> notActiveExceptionSupplier) {
             this.notActiveExceptionSupplier = notActiveExceptionSupplier;
             return self();
+        }
+
+        /**
+         * Sets whether the serialization service should (de)serialize in the
+         * compatibility (3.x) format.
+         *
+         * @param isCompatibility {@code true} if the serialized format should conform to the
+         *                        3.x serialization format, {@code false} otherwise
+         * @return this builder
+         */
+        public final T withCompatibility(boolean isCompatibility) {
+            this.isCompatibility = isCompatibility;
+            return self();
+        }
+
+        /**
+         * @return {@code true} if the serialized format of the serialization service should
+         * conform to the 3.x serialization format, {@code false} otherwise.
+         */
+        public boolean isCompatibility() {
+            return isCompatibility;
         }
     }
 }
